Remove Atlassian dependencies from beans
First of all, all Atlassian types have been removed from our model beans and
replaced by non-Atlassian types (e.g. simply String).

Constructor- / from- / to- methods have been moved to new bean util classes.
These classes will be moved to the plugin(s) with one of the next PR's so that
the Atlassian dependencies are not present in the beans anymore.

The affected bean tests have been moved to corresponding bean util tests. As the
amount of necessary refactorings was quite high and the old bean test classes
still exist, Git was not able to track these code shifts and only shows them as
"modified" and "new".

This is the last PR that is necessary before all classes that NEED to contain
Atlassian dependencies can be moved to the plugin(s). These are:

* src/main/java/de/aservo/atlassian/confapi/model/type/DefaultAuthenticationScenario
* src/main/java/de/aservo/atlassian/confapi/model/util/
* src/main/java/de/aservo/atlassian/confapi/service/ApplicationLinkServiceImpl
* src/main/java/de/aservo/atlassian/confapi/service/DirectoryServiceImpl
* src/main/java/de/aservo/atlassian/confapi/util/MailProtocolUtil

And for the tests:

* src/test/java/com/atlassian/
* src/test/java/de/aservo/atlassian/confapi/model/type/
* src/test/java/de/aservo/atlassian/confapi/model/util/
* src/test/java/de/aservo/atlassian/confapi/service/

As well as the Atlassian dependencies in the pom.xml
