+++ pom.xml
@@ -196,6 +196,13 @@
             <scope>provided</scope>
         </dependency>
 
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>1.7.30</version>
+            <scope>provided</scope>
+        </dependency>
+
         <!-- needs to be of scope 'provided' as implicitly used for 'checkNotNull' -->
         <dependency>
             <groupId>com.google.guava</groupId>
