+++ src/main/java/de/aservo/atlassian/confapi/model/DirectoryBean.java
@@ -1,8 +1,5 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.crowd.embedded.api.Directory;
-import com.atlassian.crowd.embedded.api.DirectoryType;
-import com.atlassian.crowd.model.directory.ImmutableDirectory;
 import de.aservo.atlassian.confapi.constants.ConfAPI;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,13 +8,6 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.util.HashMap;
-import java.util.Map;
-
-import static com.atlassian.crowd.directory.RemoteCrowdDirectory.*;
-import static com.atlassian.crowd.directory.SynchronisableDirectoryProperties.*;
-import static com.atlassian.crowd.directory.SynchronisableDirectoryProperties.SyncGroupMembershipsAfterAuth.WHEN_AUTHENTICATION_CREATED_THE_USER;
-import static com.atlassian.crowd.model.directory.DirectoryImpl.ATTRIBUTE_KEY_USE_NESTED_GROUPS;
 
 /**
  * Bean for user directory settings in REST requests.
@@ -42,7 +32,7 @@ public class DirectoryBean {
 
     @XmlElement
     @NotNull
-    private DirectoryType type;
+    private String type;
 
     @XmlElement
     private String description;
@@ -74,50 +64,6 @@ public class DirectoryBean {
     @XmlElement
     private String proxyPassword;
 
-    /**
-     * Build directory directory.
-     *
-     * @return the directory
-     */
-    public Directory toDirectory() {
-        Map<String, String> attributes = new HashMap<>();
-        attributes.put(CROWD_SERVER_URL, crowdUrl);
-        attributes.put(APPLICATION_PASSWORD, appPassword);
-        attributes.put(APPLICATION_NAME, clientName);
-        attributes.put(CROWD_HTTP_PROXY_HOST, proxyHost);
-        attributes.put(CROWD_HTTP_PROXY_PORT, proxyPort);
-        attributes.put(CROWD_HTTP_PROXY_USERNAME, proxyUsername);
-        attributes.put(CROWD_HTTP_PROXY_PASSWORD, proxyPassword);
-        attributes.put(CACHE_SYNCHRONISE_INTERVAL, "3600");
-        attributes.put(ATTRIBUTE_KEY_USE_NESTED_GROUPS, "false");
-        attributes.put(INCREMENTAL_SYNC_ENABLED, "true");
-        attributes.put(SYNC_GROUP_MEMBERSHIP_AFTER_SUCCESSFUL_USER_AUTH_ENABLED, WHEN_AUTHENTICATION_CREATED_THE_USER.getValue());
-
-        return ImmutableDirectory.builder(name, type, implClass).setActive(active).setAttributes(attributes).build();
-    }
-
-    /**
-     * Build user directory bean user directory bean.
-     *
-     * @param directory the directory
-     * @return the user directory bean
-     */
-    public static DirectoryBean from(Directory directory) {
-        Map<String, String> attributes = directory.getAttributes();
-        DirectoryBean directoryBean = new DirectoryBean();
-        directoryBean.setName(directory.getName());
-        directoryBean.setActive(directory.isActive());
-        directoryBean.setType(directory.getType());
-        directoryBean.setDescription(directory.getDescription());
-        directoryBean.setCrowdUrl(attributes.get(CROWD_SERVER_URL));
-        directoryBean.setClientName(attributes.get(APPLICATION_NAME));
-        directoryBean.setProxyHost(attributes.get(CROWD_HTTP_PROXY_HOST));
-        directoryBean.setProxyPort(attributes.get(CROWD_HTTP_PROXY_PORT));
-        directoryBean.setProxyUsername(attributes.get(CROWD_HTTP_PROXY_USERNAME));
-        directoryBean.setImplClass(directory.getImplementationClass());
-        return directoryBean;
-    }
-
     // Example instances for documentation and tests
 
     public static final DirectoryBean EXAMPLE_1;
@@ -127,7 +73,7 @@ public class DirectoryBean {
         EXAMPLE_1.setName("example");
         EXAMPLE_1.setCrowdUrl("https://crowd.example.com");
         EXAMPLE_1.setAppPassword("p455w0rd");
-        // EXAMPLE_1.setType("crowd");
+        EXAMPLE_1.setType("crowd");
         EXAMPLE_1.setClientName("confluence-client");
         EXAMPLE_1.setImplClass("test.class");
     }
@@ -139,7 +85,7 @@ public class DirectoryBean {
         EXAMPLE_2.setName("example");
         EXAMPLE_2.setCrowdUrl("https://localhost/crowd");
         EXAMPLE_1.setAppPassword("0th3r");
-        // EXAMPLE_2.setType("crowd");
+        EXAMPLE_2.setType("crowd");
         EXAMPLE_2.setClientName("confluence-client");
         EXAMPLE_2.setImplClass("test.class");
     }
