+++ src/main/java/de/aservo/atlassian/confapi/model/MailServerSmtpBean.java
@@ -1,12 +1,10 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.mail.server.SMTPMailServer;
 import de.aservo.atlassian.confapi.constants.ConfAPI;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import javax.annotation.Nullable;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
@@ -46,28 +44,6 @@ public class MailServerSmtpBean extends AbstractMailServerProtocolBean {
         setFrom(from);
         setPrefix(prefix);
         setHost(host);
-    }
-
-    @Nullable
-    public static MailServerSmtpBean from(
-            @Nullable final SMTPMailServer smtpMailServer) {
-
-        if (smtpMailServer == null) {
-            return null;
-        }
-
-        final MailServerSmtpBean mailServerSmtpBean = new MailServerSmtpBean();
-        mailServerSmtpBean.setName(smtpMailServer.getName());
-        mailServerSmtpBean.setDescription(smtpMailServer.getDescription());
-        mailServerSmtpBean.setFrom(smtpMailServer.getDefaultFrom());
-        mailServerSmtpBean.setPrefix(smtpMailServer.getPrefix());
-        mailServerSmtpBean.setProtocol(smtpMailServer.getMailProtocol().getProtocol());
-        mailServerSmtpBean.setHost(smtpMailServer.getHostname());
-        mailServerSmtpBean.setPort(smtpMailServer.getPort());
-        mailServerSmtpBean.setTls(smtpMailServer.isTlsRequired());
-        mailServerSmtpBean.setTimeout(smtpMailServer.getTimeout());
-        mailServerSmtpBean.setUsername(smtpMailServer.getUsername());
-        return mailServerSmtpBean;
     }
 
     // Example instances for documentation and tests
