+++ src/test/java/de/aservo/atlassian/confapi/model/MailServerSmtpBeanTest.java
@@ -1,60 +1,9 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.mail.server.DefaultTestSmtpMailServerImpl;
-import com.atlassian.mail.server.SMTPMailServer;
 import de.aservo.atlassian.confapi.junit.AbstractBeanTest;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import static com.atlassian.mail.MailConstants.DEFAULT_TIMEOUT;
-import static org.junit.Assert.*;
-
 @RunWith(MockitoJUnitRunner.class)
 public class MailServerSmtpBeanTest extends AbstractBeanTest {
-
-    @Test
-    public void testDefaultConstructor() {
-        final MailServerSmtpBean bean = new MailServerSmtpBean();
-
-        assertNull(bean.getName());
-        assertNull(bean.getDescription());
-        assertNull(bean.getFrom());
-        assertNull(bean.getPrefix());
-        assertNull(bean.getProtocol());
-        assertNull(bean.getHost());
-        assertNull(bean.getPort());
-        assertFalse(bean.isTls());
-        assertEquals(DEFAULT_TIMEOUT, bean.getTimeout());
-        assertNull(bean.getUsername());
-        assertNull(bean.getPassword());
-    }
-
-    @Test
-    public void testFromConstructor() throws Exception {
-        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
-        final MailServerSmtpBean bean = MailServerSmtpBean.from(server);
-
-        assertEquals(server.getName(), bean.getName());
-        assertEquals(server.getDescription(), bean.getDescription());
-        assertEquals(server.getDefaultFrom(), bean.getFrom());
-        assertEquals(server.getPrefix(), bean.getPrefix());
-        assertEquals(server.getMailProtocol().getProtocol(), bean.getProtocol());
-        assertEquals(server.getHostname(), bean.getHost());
-        assertEquals(Integer.valueOf(server.getPort()), bean.getPort());
-        assertEquals(server.isTlsRequired(), bean.isTls());
-        assertEquals(server.getTimeout(), bean.getTimeout());
-        assertEquals(server.getUsername(), bean.getUsername());
-        assertNull(bean.getPassword());
-    }
-
-    @Test
-    public void testFromConstructorHideEmptyDescription() throws Exception {
-        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
-        server.setDescription("");
-        final MailServerSmtpBean bean = MailServerSmtpBean.from(server);
-
-        assertNull(bean.getDescription());
-    }
-
 }
